Replace MeetSession with Meet and add meet use cases
