+++ app/src/main/kotlin/me/nemiron/khinkalyator/core/exception/utils/FlowHandleErrors.kt
@@ -1,0 +1,11 @@
+package me.nemiron.khinkalyator.core.exception.utils
+
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
+
+// TODO: add needed exception handling
+fun <T> Flow<T>.handleErrors(): Flow<T> =
+    catch { e ->
+        when (e) {
+        }
+    }
