+++ app/src/main/kotlin/me/nemiron/khinkalyator/features/meets/details/ui/RealMeetDetailsComponent.kt
@@ -1,0 +1,8 @@
+package me.nemiron.khinkalyator.features.meets.details.ui
+
+import com.arkivanov.decompose.ComponentContext
+
+class RealMeetDetailsComponent(
+    componentContext: ComponentContext,
+    configuration: MeetDetailsComponent.Configuration
+) : MeetDetailsComponent, ComponentContext by componentContext
