+++ app/src/main/kotlin/me/nemiron/khinkalyator/features/meets/home_page/ui/RealMeetsPageComponent.kt
@@ -3,11 +3,13 @@ package me.nemiron.khinkalyator.features.meets.home_page.ui
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import com.arkivanov.decompose.ComponentContext
+import kotlinx.coroutines.flow.catch
 import me.nemiron.khinkalyator.core.utils.componentCoroutineScope
 import me.nemiron.khinkalyator.core.utils.toComposeState
-import me.nemiron.khinkalyator.features.meets.domain.Meet
-import me.nemiron.khinkalyator.features.meets.domain.MeetId
+import me.nemiron.khinkalyator.common_domain.model.MeetId
+import me.nemiron.khinkalyator.common_domain.model.Meet
 import me.nemiron.khinkalyator.features.meets.domain.ObserveMeetsUseCase
+import timber.log.Timber
 
 class RealMeetsPageComponent(
     componentContext: ComponentContext,
@@ -17,7 +19,9 @@ class RealMeetsPageComponent(
 
     private val coroutineScope = componentCoroutineScope()
 
-    private val meetsState by observeMeets().toComposeState(
+    private val meetsState by observeMeets().catch {
+        Timber.tag("MEET_TAG").d("RealMeetsPageComponent.observeMeets catch: ")
+    }.toComposeState(
         initialValue = emptyList(),
         coroutineScope = coroutineScope
     )
