--- app/src/main/kotlin/me/nemiron/khinkalyator/features/meets/session/check/ui/MeetSessionCheckUi.kt
@@ -1,24 +1,0 @@
-package me.nemiron.khinkalyator.features.meets.session.check.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import me.nemiron.khinkalyator.core.theme.KhinkalyatorTheme
-
-@Composable
-fun MeetSessionCheckUi(
-    component: MeetSessionCheckComponent,
-    modifier: Modifier = Modifier
-) {
-
-}
-
-private class PreviewMeetSessionCheckComponent : MeetSessionCheckComponent
-
-@Preview
-@Composable
-private fun MeetSessionCheckPreview() {
-    KhinkalyatorTheme {
-        MeetSessionCheckUi(PreviewMeetSessionCheckComponent())
-    }
-}
