--- app/src/main/kotlin/me/nemiron/khinkalyator/features/meets/session/details/ui/RealMeetSessionDetailsComponent.kt
@@ -1,8 +1,0 @@
-package me.nemiron.khinkalyator.features.meets.session.details.ui
-
-import com.arkivanov.decompose.ComponentContext
-
-class RealMeetSessionDetailsComponent(
-    componentContext: ComponentContext,
-    configuration: MeetSessionDetailsComponent.Configuration
-) : MeetSessionDetailsComponent, ComponentContext by componentContext
