+++ app/src/main/kotlin/me/nemiron/khinkalyator/root/ui/RootComponent.kt
@@ -3,18 +3,18 @@ package me.nemiron.khinkalyator.root.ui
 import com.arkivanov.decompose.router.stack.ChildStack
 import me.nemiron.khinkalyator.features.home.ui.HomeComponent
 import me.nemiron.khinkalyator.features.meets.create.ui.CreateMeetComponent
-import me.nemiron.khinkalyator.features.meets.session.overview.ui.MeetSessionOverviewComponent
+import me.nemiron.khinkalyator.features.meets.overview.ui.MeetOverviewComponent
 import me.nemiron.khinkalyator.features.restaraunts.restaurant_overview.ui.RestaurantOverviewComponent
 import me.nemiron.khinkalyator.root.ui.start.StartComponent
 
 interface RootComponent {
-    val childStackState: ChildStack<*, Child>
+    val childStack: ChildStack<*, Child>
 
     sealed interface Child {
         class Start(val component: StartComponent) : Child
         class Home(val component: HomeComponent) : Child
         class CreateMeet(val component: CreateMeetComponent) : Child
-        class MeetSession(val component: MeetSessionOverviewComponent) : Child
+        class Meet(val component: MeetOverviewComponent) : Child
         class Restaurant(val component: RestaurantOverviewComponent) : Child
     }
 }
