+++ app/src/main/kotlin/me/nemiron/khinkalyator/root/ui/RootUi.kt
@@ -19,7 +19,7 @@ import me.nemiron.khinkalyator.core.utils.createFakeChildStack
 import me.nemiron.khinkalyator.features.home.ui.HomeUi
 import me.nemiron.khinkalyator.features.home.ui.PreviewHomeComponent
 import me.nemiron.khinkalyator.features.meets.create.ui.CreateMeetUi
-import me.nemiron.khinkalyator.features.meets.session.overview.ui.MeetSessionOverviewUi
+import me.nemiron.khinkalyator.features.meets.overview.ui.MeetOverviewUi
 import me.nemiron.khinkalyator.features.restaraunts.restaurant_overview.ui.RestaurantOverviewUi
 import me.nemiron.khinkalyator.root.ui.start.StartUi
 
@@ -31,10 +31,10 @@ fun RootUi(
 ) {
     SystemBarColors()
     Children(
-        stack = component.childStackState,
+        stack = component.childStack,
         animation = stackAnimation { child: Child.Created<Any, RootComponent.Child>, _, _ ->
             when (child.instance) {
-                is RootComponent.Child.MeetSession, is RootComponent.Child.Restaurant -> slide()
+                is RootComponent.Child.Meet, is RootComponent.Child.Restaurant -> slide()
                 is RootComponent.Child.Start, is RootComponent.Child.Home, is RootComponent.Child.CreateMeet -> fade()
             }
         },
@@ -44,7 +44,7 @@ fun RootUi(
             is RootComponent.Child.Start -> StartUi(child.component)
             is RootComponent.Child.Home -> HomeUi(child.component)
             is RootComponent.Child.CreateMeet -> CreateMeetUi(child.component)
-            is RootComponent.Child.MeetSession -> MeetSessionOverviewUi(child.component)
+            is RootComponent.Child.Meet -> MeetOverviewUi(child.component)
             is RootComponent.Child.Restaurant -> RestaurantOverviewUi(child.component)
         }
     }
@@ -74,6 +74,6 @@ private fun RootPreview() {
 }
 
 private class PreviewRootComponent : RootComponent {
-    override val childStackState =
+    override val childStack =
         createFakeChildStack(RootComponent.Child.Home(PreviewHomeComponent()))
 }
